airflow/requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-celery>=3.3.0
apache-airflow-providers-google>=10.1.0
pandas
psycopg2-binary
pytest
pyspark
requests
sqlalchemy